Revert ":tada: ChangeInfoModal renomeado para WarningModal, Warning Modal refatorado para ser usado tanto como os modais padrões com o botão de cancelar ou confirmar assim como os modais que não possuem essa configuração, além disso, uso do Modal acima na tela de Profile e no Modal da tela de Scanner sobre o QRcode não ser válido do Icod s :tada:"
This reverts commit b94688111b58528651482c859e47b726dd3c0361.

+++ src/pages/EditProfile/index.tsx
@@ -6,7 +6,8 @@ import ButtonOn from '../../assets/images/button-on.svg'
 import ButtonOff from '../../assets/images/button-off.svg'
 import DeleteAccountIcon from '../../assets/images/Icons/Profile/delete-account-icon.svg'
 import { useAuth, User } from '../../hooks/auth'
-import { WarningModal } from '../../components/WarningModal';
+import { ChangeInfoModal } from '../../components/ChangeInfoModal';
+import { useNavigation } from '@react-navigation/native';
 import theme from '../../global/styles/theme';
 import extractNameAndSurname from '../../utils/extractNameAndSurname';
 import { useTheme } from 'styled-components';
@@ -24,6 +25,7 @@ interface EditProfileProps {
 type UserFields = 'id' | 'name' | 'email' | 'username' | 'visibility'
 
 const EditProfile = ({ route }: EditProfileProps) => {
+  const navigation = useNavigation();
   const theme = useTheme();
   const { user, token, alterProfileVisibility, signOut, deleteUser } = useAuth()
   const [deleteAccountModalOpen, setDeleteAccountModalOpen] = useState(false)
@@ -64,15 +66,14 @@ const EditProfile = ({ route }: EditProfileProps) => {
         edit
         avatar={avatar}
       />
-      <WarningModal
+      <ChangeInfoModal
         title={'Você está prestes a excluir a conta'}
-        description={'Ao confirmar, seus dados serão excluidos e será necessário \n fazer outro cadastro'}
-        icon={DeleteAccountIcon}
-        isFooterButtonsActived
-        iconBackgroundColor={theme.colors.attention}
+        description={'Ao confirmar, seus dados serão excluidos e será necessário fazer outro cadastro'}
+        icon={<DeleteAccountIcon />}
+        iconBackgroundColor={'#ce3e36'}
         visible={deleteAccountModalOpen}
         pressedOut={() => setDeleteAccountModalOpen(!deleteAccountModalOpen)}
-        handleSaveUpdatesconfirmed={async () => {
+        confirmed={async () => {
           setDeleteAccountModalOpen(false)
           await signOut()
           await deleteUser(token)
